backend/pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "texgrep-backend"
version = "0.1.0"
description = "Texgrep Django backend"
authors = [{ name = "texgrep" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "Django>=5.0,<6.0",
    "djangorestframework>=3.15",
    "django-cors-headers>=4.3",
    "uvicorn[standard]>=0.29",
    "celery[redis]>=5.3",
    "redis>=5.0",
    "httpx>=0.27",
    "pydantic>=2.6",
    "opensearch-py>=2.4",
    "python-dotenv>=1.0",
    "structlog>=24.1",
    "python-dateutil>=2.8",
    "django-ratelimit>=4.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1",
    "pytest-django>=4.8",
    "pytest-asyncio>=0.23",
    "factory_boy>=3.3",
    "freezegun>=1.4",
    "ruff>=0.3",
]

[tool.setuptools]
packages = ["texgrep", "search"]

[tool.black]
line-length = 140
target-version = ["py310"]
extend-exclude = '''
(
  ^\.venv/
| ^build/
| ^dist/
| ^.*/migrations/
)
'''

[tool.ruff]
target-version = "py310"
line-length = 140

[tool.ruff.lint]
select = ["E","F","I","W","UP","B","ISC","PL","SIM","RUF"]
ignore = ["RUF003","E501"]   # 全角カッコ／長行は黙認

[tool.ruff.lint.per-file-ignores]
# pyproject が backend/ にあるので、ここからの相対パスで指定する
"search/tests/**" = ["PLR2004","PLC0415","RUF003"]
"search/providers/__init__.py" = ["PLC0415"]   # 動的 import を許容するなら
"manage.py" = ["PLC0415"]
"search/query.py" = ["PLR0912"]                 # まずは抑制（余裕があれば後で関数分割）
"search/snippets.py" = ["PLR0912","PLR0915"]    # 同上

[tool.ruff.format]
quote-style = "preserve"

[tool.ruff.lint.isort]
# Black 互換の設定（デフォでOKだが明示）
combine-as-imports = true
force-single-line = false

[tool.mypy]
python_version = "3.10"
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
mypy_path = ["backend"]
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
strict_optional = true
# 段階的に上げる:
disallow_untyped_defs = false
follow_imports = "silent"

[tool.django-stubs]
django_settings_module = "texgrep.settings"

[tool.coverage.run]
branch = true
source = [
  "backend/search",
  "indexer",
]
omit = [
  # 実行時に関与しないスクリプト/エントリポイント
  "backend/manage.py",
  "backend/texgrep/asgi.py",
  "backend/texgrep/wsgi.py",
  "backend/search/management/*",
  "indexer/main.py",
  # マイグレーション等
  "*/migrations/*",
]

[tool.coverage.report]
show_missing = true
skip_covered = true
# まずは現実的なラインに。対象を絞ったら徐々に引き上げるのが◎
fail_under = 80

# あるある除外（型用/未実装等）
exclude_also = [
  "pragma: no cover",
  "if TYPE_CHECKING:",
  "raise NotImplementedError",
  "def __repr__",
]
